+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.25.0] - Not released
+### Fixed
+- Fix attachment sanitization when the file isn't found on S3 storage.
 
 ## [2.24.0] - 2025-06-24
 ### Fixed